requirements.txt
Flask
requests
yt-dlp
youtube-transcript-api
google-generativeai
gunicorn
python-dotenv
rq
